requirements.txt
bentoml
pillow
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
requests==2.31.0
